requirements.txt
requests==2.32.5
beautifulsoup4==4.14.3
gspread==6.2.1
google-auth==2.49.0